added basic combat system, added placeholder audio, added shitty skybox, basic audio management, some other stuff

+++ Assets/InventoryMaster/Scripts/Item/ItemAttribute.cs
@@ -6,8 +6,8 @@ public class ItemAttribute
 {
 
     public string attributeName;
-    public int attributeValue;
-    public ItemAttribute(string attributeName, int attributeValue)
+    public float attributeValue;
+    public ItemAttribute(string attributeName, float attributeValue)
     {
         this.attributeName = attributeName;
         this.attributeValue = attributeValue;
